对f072的bootloader接收can数据解析修改

+++ STM32F072C8T6_IAP/bootloader/HARDWARE/can.h
@@ -8,7 +8,7 @@ typedef enum {FAILED = 0, PASSED = !FAILED} TestStatus;
 void CAN_Configuration(uint32_t BaudRate);
 uint8_t CAN_WriteData(CanTxMsg *TxMessage);
 void BOOT_DelayMs(uint32_t ms);
-uint16_t Read_CAN_Address(void);
+uint8_t Read_CAN_Address(void);
 TestStatus CAN_Polling(void);
 TestStatus CAN_Interrupt(void);
